added all corner cases tests

+++ src/main/java/com/horiaconstantin/distanceCalculator/service/Distance.java
@@ -1,10 +1,15 @@
 package com.horiaconstantin.distanceCalculator.service;
 
+import lombok.NonNull;
 import lombok.Value;
 
+/**
+ * An object representing a distance.
+ * Constaints: the value must be positive and the unit of measurement must be defined
+ */
 @Value
 class Distance {
-    private int value;
+    private double value;
     private Unit unit;
 
 }
